feat: add basic and book blueprints with endpoints for health, course, custom response, and book lookup
- Implement /health, /course, and /custom endpoints in basic blueprint

- Implement dynamic /book/<isbn> route in book blueprint using Open Library API

- Prepare app for modular route handling with Flask blueprints
